# api/swagger.yaml
openapi: 3.0.0
info:
  title: Auth Service API
  description: Authentication service API with email/password and Google OAuth support
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.example.com
    description: Production server

tags:
  - name: Authentication
    description: Authentication endpoints

paths:
  /api/v1/auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new user account with email and password
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
            example:
              email: user@example.com
              password: password123
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '400':
          description: Bad request - Invalid input or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "invalid request body"

  /api/v1/auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticate a user with email and password
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginUserRequest'
            example:
              email: user@example.com
              password: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '400':
          description: Bad request - Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "invalid request body"
        '401':
          description: Unauthorized - Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "invalid credentials"

  /api/v1/auth/login/google:
    post:
      tags:
        - Authentication
      summary: Login with Google OAuth
      description: Authenticate a user using Google OAuth ID token
      operationId: loginWithGoogle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginWithGoogleRequest'
            example:
              id_token: eyJhbGciOiJSUzI1NiIsImtpZCI6IjEyMzQ1NiJ9...
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '400':
          description: Bad request - Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "invalid request body"
        '401':
          description: Unauthorized - Invalid or expired Google token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "invalid google token"

  /api/v1/auth/google/challenge:
    get:
      tags:
        - Authentication
      summary: Get Google OAuth challenge URL
      description: Retrieve the Google OAuth authorization URL for initiating the OAuth flow
      operationId: getGoogleChallenge
      responses:
        '200':
          description: Google OAuth URL retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoogleOAuthChallengeResponse'
              example:
                auth_url: https://accounts.google.com/o/oauth2/v2/auth?client_id=...
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    CreateUserRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 8
          description: User password (minimum 8 characters)
          example: password123

    LoginUserRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: user@example.com
        password:
          type: string
          format: password
          description: User password
          example: password123

    LoginWithGoogleRequest:
      type: object
      required:
        - id_token
      properties:
        id_token:
          type: string
          description: Google OAuth ID token
          example: eyJhbGciOiJSUzI1NiIsImtpZCI6IjEyMzQ1NiJ9...

    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT authentication token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...

    GoogleOAuthChallengeResponse:
      type: object
      properties:
        auth_url:
          type: string
          format: uri
          description: Google OAuth authorization URL
          example: https://accounts.google.com/o/oauth2/v2/auth?client_id=...

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: An error occurred